+++ src/Hedwig/ClientApp/src/containers/Enrollment/Edit/EnrollmentEdit.tsx
@@ -3,7 +3,7 @@ import { Section } from '../enrollmentTypes';
 import { History } from 'history';
 import ChildInfo from '../_sections/ChildInfo';
 import FamilyInfo from '../_sections/FamilyInfo';
-import FamilyIncome from '../_sections/FamilyIncome_New';
+import FamilyIncome from '../_sections/FamilyIncome';
 import EnrollmentFunding from '../_sections/EnrollmentFunding';
 import PageNotFound from '../../PageNotFound/PageNotFound';
 import UserContext from '../../../contexts/User/UserContext';
