+++ src/Hedwig/ClientApp/src/containers/Enrollment/_sections/FamilyIncome/UpdateForm/DeterminationCard/CardContentWithExpandCard.tsx
